Implementing Tag and Task classes
Basic implementation for the Tag and the Task class. I did this so that the other classes can be worked on without needing the use of blackbox functions or any other workarounds. Left out some functions like the deleteTask() function, so that the people who are responsible for this component can implement.