Worked on UI
Added a taskbar at the top to always display "My Tasks" and the current date. Add a tab area and made it scrollable for when we add tags to it later. Created a action button that is always present at the bottom right of the screen that will be connected to our create_Task function.